txn: destroy commit and rollback triggers
They were not deleted ever. Worked fine for DDL, for which they
were introduced on the first place, because these triggers are on
the region memory.

But didn't work when the triggers became available in the public
API, because these are allocated on the heap. As a result, all the
box.on_commit() and box.on_rollback() triggers leaked.

The patch ensures all the on_commit/on_rollback triggers are
destroyed. In order to mitigate the performance impact the
internal triggers now clear themselves so as it wouldn't be
necessary to walk their list in txn.c again.

Another hack not to touch these triggers when possible makes their
destruction happen only in txn_run_commit_triggers() and
txn_run_rollback_triggers(). These are not called when a
transaction does not have any triggers, and therefore these
destroys don't cost anything. On the contrary with if they would
be destroyed in txn_free() which is called always.

In short: the patch does not affect the common path when the
triggers are not used.

Another option was to force all the triggers clear themselves. For
example, in case of commit all the on_commit triggers must clear
themselves, and the rollback triggers are destroyed. Vice versa
when a rollback happens. But it didn't work because the Lua
triggers all work via a common runner lbox_trigger_run(), which
can't destroy its argument in most of the cases (space:on_replace,
:before_replace, ...).

Closes #6025